+++ Documentation~/version-history/whats-new.md
@@ -1,7 +1,32 @@
 ---
 uid: xrhands-whats-new
 ---
-# What's new in version 1.0
+# What's new in version 1.2
+
+Summary of changes in XR Hands package version 1.2.
+
+The main updates in this release include:
+
+**Added**
+- Added `IXRHandProcessor` interface that intercepts and allows post-processing joint data before being used by other scripts. More information can be found on the [XR Hands Joint Processing](../process-joints.md) manual page.
+- Added two samples to the Hand Visualizer sample for how to use the new Joint Processing feature to change the visual output of hand tracking.
+- `XRHand` and `XRHandJoint` both now implement `IEquatable`, and have `Equals`, `GetHashCode`, and the `==` and `!=` operators overloaded.
+- Added `XRHandTrackingEvents` component that subscribes to hand tracking events for a specific hand and surfaces Unity Events when the hand updates or tracking state changes.
+- Added `XRHandMeshController` component that subscribes to events from a `XRHandTrackingEvents` and can optionally show or hide a mesh.
+- Added `XRHandSkeletonDriver` component that subscribes to events from a `XRHandTrackingEvents` and updates a hierarchy of Transforms for each `XRHandJoint`.
+
+**Changed**
+- Minimum supported Unity Editor version has changed to 2021.3.
+
+**Fixed**
+
+- Fixed Hand Visualizer component so it will use another running `XRHandSubsystem` if the first is stopped.
+- Fixed exception in Hand Visualizer component when a joint is invalid.
+- Fixed rendering of lines that connect joints in `HandVisualizer` sample when used with the device simulator. Also removed the Origin property from the component since it was no longer used.
+
+
+---
+# What was new in version 1.0
 
 Summary of changes in XR Hands package version 1.0.